requirements.txt
sentence-transformers
transformers
torch
python-dotenv